+++ libraries/resource_chef_dk_app_debian.rb
@@ -72,13 +72,23 @@ class Chef
 
       #
       # Upgrade or install the Chef-DK. This action currently only supports the
-      # :repo installation source.
+      # :direct and :repo installation sources.
       #
       action :upgrade do
         case new_resource.source
         when :direct
-          raise(Chef::Exceptions::UnsupportedAction,
-                'Direct installs do not support the :upgrade action')
+          new_resource.installed(true)
+          new_resource.version('latest')
+
+          converge_if_changed :installed, :version do
+            local_path = ::File.join(Chef::Config[:file_cache_path],
+                                     ::File.basename(package_metadata[:url]))
+            remote_file local_path do
+              source package_metadata[:url]
+              checksum package_metadata[:sha256]
+            end
+            dpkg_package local_path
+          end
         when :repo
           package 'apt-transport-https'
           include_recipe "apt-chef::#{new_resource.channel}"